refactor(server): Split in modules
